voice/requirements.txt
google-cloud-firestore==0.32.0
Flask==1.0.2
google-auth==1.6.3